added ATTiny1614 code

+++ src/filesystem.h
@@ -17,8 +17,8 @@ bool initFilesystem();
 
 // y = x + p/100 * (x^3 - x)
 struct ExponentAndMax{
-  int8_t exponent; /** Value from -100 to 100. if negative, we just reflect the function over y=x (point a;b becomes b;a). */
-  uint8_t maxValue {100}; /** Value 0 - 100, 100 by default */
+  int exponent; /** Value from -100 to 100. if negative, we just reflect the function over y=x (point a;b becomes b;a). */
+  int maxValue {100}; /** Value 0 - 100, 100 by default */
 };
 
 struct Model{
@@ -29,12 +29,12 @@ struct Model{
   bool reversedSwitches2[4]; /** true if reversed */
   bool reversedSwitches3[4]; /** true if reversed */
 
-  int8_t flightModeSwitch2; /** -1 if no fly mode switch is assigned. Only 2-state switches can be assigned. */
-  int8_t throttleHoldSwitch2; /** -1 if no throttle hold switch is assigned. Only 2-state switches can be assigned. */
+  int flightModeSwitch2; /** -1 if no fly mode switch is assigned. Only 2-state switches can be assigned. */
+  int throttleHoldSwitch2; /** -1 if no throttle hold switch is assigned. Only 2-state switches can be assigned. */
   ExponentAndMax flightMode1[6];
   ExponentAndMax flightMode2[6];
 
-  int8_t subtrim[6]; /** values -100 to +100, correcting the input of each variable channel */
+  int subtrim[6]; /** values -100 to +100, correcting the input of each variable channel */
 };
 
 /**
